requirements.txt
aiogram==3.10.0
python-dotenv==1.0.0
gspread==5.12.0
google-auth==2.25.0